refactor(ADFA-1653): Move feedback fab logic to separate class
